Move the invitation code to a route
